fix formatting

+++ contracts/contracts-foundry/test/kms/KmsVerifiable.t.sol
@@ -5,9 +5,8 @@ import {Test, console} from "../../lib/forge-std/src/Test.sol";
 import {KmsVerifiable} from "../../src/kms/KmsVerifiable.sol";
 
 contract KmsVerifiableTestConstructor is Test {
-    function test_construction(
-    ) public {
-        bytes32 [] memory _imageIds = new bytes32[](1);
+    function test_construction() public {
+        bytes32[] memory _imageIds = new bytes32[](1);
         address _owner = makeAddr("owner");
         vm.prank(_owner);
         _imageIds[0] = bytes32(vm.randomUint());
@@ -23,6 +22,7 @@ contract KmsVerifiableTest is Test {
     address owner;
     bytes32[] imageIds;
     KmsVerifiable kmsVerifiable;
+
     function setUp() public {
         owner = makeAddr("owner");
         vm.prank(owner);
@@ -42,24 +42,24 @@ contract KmsVerifiableTest is Test {
     function test_approveImages(bytes32[] calldata _imageIds) public {
         // Execute
         vm.prank(owner);
-        for (uint i = 0; i < _imageIds.length; i++) {
+        for (uint256 i = 0; i < _imageIds.length; i++) {
             vm.expectEmit();
             emit KmsVerifiable.ImageApproved(_imageIds[i]);
         }
         kmsVerifiable.approveImages(_imageIds);
         // Validate
-        for (uint i = 0; i < _imageIds.length; i++) {
+        for (uint256 i = 0; i < _imageIds.length; i++) {
             assertTrue(kmsVerifiable.images(_imageIds[i]));
         }
 
         // approve the same images again
         vm.prank(owner);
-        for (uint i = 0; i < _imageIds.length; i++) {
+        for (uint256 i = 0; i < _imageIds.length; i++) {
             vm.expectEmit();
             emit KmsVerifiable.ImageApproved(_imageIds[i]);
         }
         kmsVerifiable.approveImages(_imageIds);
-        for (uint i = 0; i < _imageIds.length; i++) {
+        for (uint256 i = 0; i < _imageIds.length; i++) {
             assertTrue(kmsVerifiable.images(_imageIds[i]));
         }
     }
@@ -67,24 +67,24 @@ contract KmsVerifiableTest is Test {
     function test_revokeImages() public {
         // Execute
         vm.prank(owner);
-        for (uint i = 0; i < imageIds.length; i++) {
+        for (uint256 i = 0; i < imageIds.length; i++) {
             vm.expectEmit();
             emit KmsVerifiable.ImageRevoked(imageIds[i]);
         }
         kmsVerifiable.revokeImages(imageIds);
         // Validate
-        for (uint i = 0; i < imageIds.length; i++) {
+        for (uint256 i = 0; i < imageIds.length; i++) {
             assertFalse(kmsVerifiable.images(imageIds[i]));
         }
 
         // revoke the same images again
         vm.prank(owner);
-        for (uint i = 0; i < imageIds.length; i++) {
+        for (uint256 i = 0; i < imageIds.length; i++) {
             vm.expectEmit();
             emit KmsVerifiable.ImageRevoked(imageIds[i]);
         }
         kmsVerifiable.revokeImages(imageIds);
-        for (uint i = 0; i < imageIds.length; i++) {
+        for (uint256 i = 0; i < imageIds.length; i++) {
             assertFalse(kmsVerifiable.images(imageIds[i]));
         }
     }
